publish jdcloud-sdk-go version 1.0.3

+++ services/waf/apis/GetBpsData.go
@@ -112,6 +112,6 @@ type GetBpsDataResponse struct {
 }
 
 type GetBpsDataResult struct {
-    TimeScope []int `json:"timeScope"`
+    TimeScope []int64 `json:"timeScope"`
     Bps waf.Bps `json:"bps"`
 }
